% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_bayes_classifier.R
\name{naive_bayes_classifier}
\alias{naive_bayes_classifier}
\title{NaiveBayesClassifier}
\description{
This is a package that implements in R6Class the naive bayes classifier algorithme with parallel execution
the available types for now are gaussian, default, multinomial, binomial
}
\examples{
#obj= naive_bayes_classifier$new()

## ------------------------------------------------
## Method `naive_bayes_classifier$new`
## ------------------------------------------------

#obj= naive_bayes_classifier$new(type="default")


## ------------------------------------------------
## Method `naive_bayes_classifier$fit`
## ------------------------------------------------

#X= iris[,1:4]
#y= iris[-,5]
#obj$fit(X, y)

## ------------------------------------------------
## Method `naive_bayes_classifier$predict_proba`
## ------------------------------------------------

#obj$predict_prob(X)

## ------------------------------------------------
## Method `naive_bayes_classifier$predict`
## ------------------------------------------------

#obj$predict(X)


## ------------------------------------------------
## Method `naive_bayes_classifier$summary`
## ------------------------------------------------

#obj$summary()

## ------------------------------------------------
## Method `naive_bayes_classifier$print`
## ------------------------------------------------

#obj$print()
#print(obj)

## ------------------------------------------------
## Method `naive_bayes_classifier$evaluate`
## ------------------------------------------------

#ypred= obj$predict(X)
#ytrue= y
#obj$evaluate(ypred= ypred, ytrue=ytrue, positive_class='Setosa')
#obj$evaluate(xtest= X, ytrue=ytrue, positive_class='Setosa')

## ------------------------------------------------
## Method `naive_bayes_classifier$components_plot`
## ------------------------------------------------

#obj$components_plot()
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{getModel}}{Return the getter of conditionnal probabilites of X knowing y}

\item{\code{getPrior}}{Return the getter of priors y informations}

\item{\code{getClasses}}{Return the getter of Y levels
naive_bayes_classifier R6 init class}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NaiveBayesClassifier-new}{\code{naive_bayes_classifier$new()}}
\item \href{#method-NaiveBayesClassifier-fit}{\code{naive_bayes_classifier$fit()}}
\item \href{#method-NaiveBayesClassifier-predict_proba}{\code{naive_bayes_classifier$predict_proba()}}
\item \href{#method-NaiveBayesClassifier-predict}{\code{naive_bayes_classifier$predict()}}
\item \href{#method-NaiveBayesClassifier-summary}{\code{naive_bayes_classifier$summary()}}
\item \href{#method-NaiveBayesClassifier-print}{\code{naive_bayes_classifier$print()}}
\item \href{#method-NaiveBayesClassifier-evaluate}{\code{naive_bayes_classifier$evaluate()}}
\item \href{#method-NaiveBayesClassifier-components_plot}{\code{naive_bayes_classifier$components_plot()}}
\item \href{#method-NaiveBayesClassifier-clone}{\code{naive_bayes_classifier$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayesClassifier-new"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayesClassifier-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{naive_bayes_classifier$new(
  type = "default",
  smoothness = 1,
  priors = NULL,
  reduction = NULL,
  n_components = 2,
  discretize = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{the type of naive algo you want to use (must be default, gaussian, bernoulli, multinomial or complement )}

\item{\code{smoothness}}{is set to avoid 0 probability effect (if 1 it is laplace)}

\item{\code{priors}}{null by default (it's the priors probabilities of Y)}

\item{\code{reduction}}{null by default (can be pca, mca or famd to apply reduction analysis on our data)}

\item{\code{n_components}}{2 by default if reduction is set. To control your the number of components you keep for your fit}

\item{\code{discretize}}{if you'd like to discretize your data. (from numeric to categoricals clusters) can be null, non-gaussian(to discretize non gaussian value by checking if normal with shapiro), all}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing (but will prepare the class to execute some tasks as fit, predict, and so on...)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{#obj= naive_bayes_classifier$new(type="default")

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayesClassifier-fit"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayesClassifier-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{naive_bayes_classifier$fit(X, y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{dataframe/matrix object with data to form the model on.}

\item{\code{y}}{the classes corresponding to X (must be a vector of characters with the same length as the number of X rows)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
will save the model (condtional probabilities of X by y) and the prior (y classes frequency)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{#X= iris[,1:4]
#y= iris[-,5]
#obj$fit(X, y)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayesClassifier-predict_proba"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayesClassifier-predict_proba}{}}}
\subsection{Method \code{predict_proba()}}{
to predict the probabilities of new data given to the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{naive_bayes_classifier$predict_proba(newdata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newdata}}{must be of dataframe type}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the probabilities conditionnaly to y
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{#obj$predict_prob(X)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayesClassifier-predict"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayesClassifier-predict}{}}}
\subsection{Method \code{predict()}}{
will use conditional probabilities and select the classes of those wo had the biggest probabilities
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{naive_bayes_classifier$predict(newdata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newdata}}{must be of dataframe type}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the class of each row of your dataframe (it will just take the class of the biggest proba)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{#obj$predict(X)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayesClassifier-summary"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayesClassifier-summary}{}}}
\subsection{Method \code{summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{naive_bayes_classifier$summary(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the result of class object after doing the fit. (you can see the type, priors probabilitites, the conditional ones, )
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{#obj$summary()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayesClassifier-print"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayesClassifier-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{naive_bayes_classifier$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
some details informations about your class object
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{#obj$print()
#print(obj)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayesClassifier-evaluate"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayesClassifier-evaluate}{}}}
\subsection{Method \code{evaluate()}}{
(either you give it xtest and ytrue and it will do the prediction and evaluate it on ytrue or you give ytest and ytrue and it will give the evaluation)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{naive_bayes_classifier$evaluate(
  ypred = NULL,
  ytrue = NULL,
  xtest = NULL,
  positive_class = "default"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ypred}}{the classes you predicted on a dataframe (result of predict function) (optionnal if you pass xtest)}

\item{\code{ytrue}}{the true classes of the same dataframe you uses to have ypred (xtest)}

\item{\code{xtest}}{the data to make the prediction on. (optional if you pass ypred already)}

\item{\code{positive_class}}{give the classe that you'd like to make the focus on. (if not specified will act by default)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
evaluation of confusion matrix on the given data
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{#ypred= obj$predict(X)
#ytrue= y
#obj$evaluate(ypred= ypred, ytrue=ytrue, positive_class='Setosa')
#obj$evaluate(xtest= X, ytrue=ytrue, positive_class='Setosa')
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayesClassifier-components_plot"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayesClassifier-components_plot}{}}}
\subsection{Method \code{components_plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{naive_bayes_classifier$components_plot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the plot of pca or mca or famd components variables results
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{#obj$components_plot()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NaiveBayesClassifier-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NaiveBayesClassifier-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{naive_bayes_classifier$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
