% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{patients}
\alias{patients}
\title{Patients}
\format{
A data frame with 17 columns and 1500 rows variables:
\describe{
\item{\code{Patient ID}}{patient id}
\item{\code{Age}}{age}
\item{\code{Gender}}{patient gender}
\item{\code{Ethnicity}}{his ethnicity}
\item{\code{Marital Status}}{his marital status}
\item{\code{Education Level}}{his education level}
\item{\code{OCD Diagnosis Date}}{his ocd diagnosis date}
\item{\code{Duration of Symptoms (months)}}{symptome duration by month}
\item{\code{Previous Diagnoses}}{his previous diagnosis}
\item{\code{Family History of OCD}}{OCDs in his family?}
\item{\code{Obsession Type}}{type of the patient obsession}
\item{\code{Compulsion Type}}{the compulsion type}
\item{\code{Y-BOCS Score (Obsessions)}}{score obsession}
\item{\code{Y-BOCS Score (Compulsions)}}{compulsion score}
\item{\code{Depression Diagnosis}}{the diagnosis of the depression}
\item{\code{Anxiety Diagnosis}}{his anxity diagnosis}
\item{\code{Medications}}{patient medication}
}
}
\usage{
patients
}
\description{
OCD diagnosis on some patients (dataframe file)
}
\keyword{datasets}
