% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{students}
\alias{students}
\title{Students}
\format{
A data frame with 34 columns and 1500 rows
}
\usage{
students
}
\description{
some students from Portugal informations. Please load it to see more informations
}
\keyword{datasets}
